requirements.txt
fastapi
uvicorn
celery
redis
google-cloud-storage
python-dotenv
